backend/pyproject.toml
[project]
name = "omni-backend"
version = "0.1.0"
description = "Omni — Multi-client AI agent hub backend (FastAPI + ADK + Raw WebSocket)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # --- Core AI ---
    "google-adk>=0.5.0",
    "google-genai>=1.0.0",
    # --- Web Framework ---
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=14.0",
    # --- Firebase / GCP ---
    "firebase-admin>=6.5.0",
    "google-cloud-firestore>=2.16.0",
    "google-cloud-storage>=2.16.0",
    "google-cloud-secret-manager>=2.20.0",
    "google-cloud-logging>=3.10.0",
    "google-cloud-trace>=1.12.0",
    "google-cloud-aiplatform[agent_engines]>=1.112.0",
    "google-cloud-scheduler>=2.13.0",
    "google-auth>=2.29.0",
    # --- E2B Sandbox ---
    "e2b-code-interpreter>=1.0.0",
    "e2b-desktop>=2.3.0",
    # --- Utilities ---
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "structlog>=24.0.0",
    "wikipedia>=1.4.0",
    "croniter>=2.0.0",
    "stripe>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
